+++ main.go
@@ -2,11 +2,8 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"html"
-	"io"
 	"log/slog"
 	"net"
 	"net/http"
@@ -15,17 +12,13 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/go-ldap/ldap/v3"
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 
-	slogmulti "github.com/samber/slog-multi"
-
 	"389-ds-exporter/src/cmd"
-	"389-ds-exporter/src/collectors"
 	"389-ds-exporter/src/config"
 	"389-ds-exporter/src/connections"
 	"389-ds-exporter/src/metrics"
+	"389-ds-exporter/src/utils"
 )
 
 var (
@@ -36,7 +29,6 @@ var (
 )
 
 const (
-	LogFileMode               os.FileMode   = 0o644
 	LdapConnectionPoolTimeout time.Duration = 3 * time.Second
 )
 
@@ -48,7 +40,7 @@ type appResources struct {
 	HttpServer *http.Server
 }
 
-func (r *appResources) Shutdown(ctx context.Context) error {
+func (r *appResources) shutdown(ctx context.Context) error {
 	slog.Info("Shutting down gracefully...")
 
 	var errs []error
@@ -78,158 +70,6 @@ func (r *appResources) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-// defaultHttpResponse function generates a standard HTML response for the exporter.
-func defaultHttpResponse(metricsPath string) func(w http.ResponseWriter, r *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		_, err := fmt.Fprintf(w, `<html>
-	<head>
-		<title>389-ds-exporter</title>
-	</head>
-	<body>
-		<p>Metrics are <a href="%s">here</a></p>
-	</body>
-</html>
-`, html.EscapeString(metricsPath))
-		if err != nil {
-			slog.Error("Error writing HTTP answer", "err", err)
-		}
-	}
-}
-
-// defaultHttpResponse function generates a standard HTML response for the exporter.
-func healthHttpResponse(pool *connections.LdapConnectionPool, startTime time.Time) func(w http.ResponseWriter, r *http.Request) {
-	return func(w http.ResponseWriter, req *http.Request) {
-
-		ldapStatus := "ok"
-		ldapAvailable := true
-		// Since the pool checks connections before issuing them,
-		// and gives out either a verified (live) connection or a newly established one,
-		// we can assume that if the pool has issued a connection, ldap is available.
-		ldapReq := ldap.NewSearchRequest(
-			"",
-			ldap.ScopeBaseObject,
-			ldap.NeverDerefAliases,
-			1, 0, false,
-			"(objectClass=*)",
-			[]string{"dn"},
-			nil,
-		)
-		ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-		defer cancel()
-		conn, err := pool.Get(ctx)
-
-		if err != nil {
-			slog.Warn("Healthcheck error", "err", err)
-			ldapStatus = "unavailable"
-			ldapAvailable = false
-		} else {
-			defer conn.Close()
-
-			_, err = conn.Search(ldapReq)
-			if err != nil {
-				slog.Warn("LDAP health check failed", "err", err)
-				ldapStatus = "unavailable"
-				ldapAvailable = false
-			}
-		}
-
-		uptime := time.Since(startTime).Seconds()
-
-		healthResponse := map[string]any{
-			"status": map[string]string{
-				"ldap": ldapStatus,
-			},
-			"uptime_seconds": int(uptime),
-			"timestamp":      time.Now().Format(time.RFC3339),
-		}
-
-		if ldapAvailable {
-			w.WriteHeader(http.StatusOK)
-		} else {
-			w.WriteHeader(http.StatusServiceUnavailable)
-		}
-
-		w.Header().Set("Content-Type", "application/json")
-		err = json.NewEncoder(w).Encode(healthResponse)
-		if err != nil {
-			slog.Error("Failed to write health response", "err", err)
-		}
-	}
-}
-
-func buildLogHandler(format string, w io.Writer, level slog.Level) slog.Handler {
-	switch format {
-	case "text":
-		return slog.NewTextHandler(w, &slog.HandlerOptions{AddSource: true, Level: level})
-	case "json":
-		return slog.NewJSONHandler(w, &slog.HandlerOptions{AddSource: true, Level: level})
-	}
-
-	return slog.Default().Handler()
-}
-
-func setupLogger(cfg *config.ExporterConfiguration) (*slog.Logger, *os.File, error) {
-	var logLevel slog.Level
-	handlers := []slog.Handler{}
-	var logFile *os.File
-
-	strLogLevel := cfg.Logging.GetLevel()
-	levelMap := map[string]slog.Level{
-		"DEBUG":   slog.LevelDebug,
-		"INFO":    slog.LevelInfo,
-		"WARNING": slog.LevelWarn,
-		"ERROR":   slog.LevelError,
-	}
-
-	logLevel, ok := levelMap[strLogLevel]
-	if !ok {
-		return nil, nil, fmt.Errorf("unknown logging level: '%s'", strLogLevel)
-	}
-
-	if cfg.Logging.GetHandler() == "stdout" || cfg.Logging.GetHandler() == "both" {
-		handlers = append(handlers, buildLogHandler(cfg.Logging.GetStdoutFormat(), os.Stdout, logLevel))
-	}
-	if cfg.Logging.GetHandler() == "file" || cfg.Logging.GetHandler() == "both" {
-		var err error
-		logFile, err = os.OpenFile(cfg.Logging.GetFile(), os.O_CREATE|os.O_WRONLY|os.O_APPEND, LogFileMode)
-		if err != nil {
-			return nil, nil, fmt.Errorf("error opening log file: %w", err)
-		}
-		handlers = append(handlers, buildLogHandler(cfg.Logging.GetFileFormat(), logFile, logLevel))
-	}
-
-	if len(handlers) == 0 {
-		return nil, nil, errors.New("unable to create logger - logging handlers not specified")
-	}
-
-	logger := slog.New(slogmulti.Fanout(handlers...))
-
-	return logger, logFile, nil
-}
-
-func reopenLogFile(cfg *config.ExporterConfiguration, resources *appResources) error {
-	if resources.LogFile != nil {
-		_ = resources.LogFile.Close()
-	}
-
-	logFile, err := os.OpenFile(cfg.Logging.GetFile(), os.O_CREATE|os.O_WRONLY|os.O_APPEND, LogFileMode)
-	if err != nil {
-		return fmt.Errorf("failed to open new log file: %w", err)
-	}
-
-	newLogger, _, err := setupLogger(cfg)
-	if err != nil {
-		_ = logFile.Close()
-
-		return fmt.Errorf("failed to set up new logger: %w", err)
-	}
-
-	slog.SetDefault(newLogger)
-	resources.LogFile = logFile
-
-	return nil
-}
-
 func readConfig(configFilePath string) (*config.ExporterConfiguration, error) {
 	configuration, err := config.ReadConfig(configFilePath)
 	if err != nil {
@@ -244,104 +84,6 @@ func readConfig(configFilePath string) (*config.ExporterConfiguration, error) {
 	return &configuration, nil
 }
 
-func setupPrometheusMetrics(
-	cfg *config.ExporterConfiguration,
-	connPool *connections.LdapConnectionPool,
-) *prometheus.Registry {
-	dsMetricsRegistry := prometheus.NewRegistry()
-
-	dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-		"ds_exporter",
-		connPool,
-		"cn=monitor",
-		metrics.GetLdapServerMetrics(),
-		prometheus.Labels{},
-		LdapConnectionPoolTimeout,
-	),
-	)
-
-	dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-		"ds_exporter",
-		connPool,
-		"cn=snmp,cn=monitor",
-		metrics.GetLdapServerSnmpMetrics(),
-		prometheus.Labels{},
-		LdapConnectionPoolTimeout,
-	),
-	)
-
-	for _, entry := range cfg.Global.NumSubordinatesRecords {
-		dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-			"ds_exporter",
-			connPool,
-			entry,
-			metrics.GetEntryCountAttr(),
-			prometheus.Labels{"entry": entry},
-			LdapConnectionPoolTimeout,
-		),
-		)
-	}
-
-	for _, backend := range cfg.Global.Backends {
-		dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-			"ds_exporter",
-			connPool,
-			"cn=monitor,cn="+backend+",cn=ldbm database,cn=plugins,cn=config",
-			metrics.GetLdapBackendCaches(),
-			prometheus.Labels{"database": backend},
-			LdapConnectionPoolTimeout,
-		),
-		)
-	}
-
-	/*
-		Since 389-ds has a different set of monitoring metrics for different backends (Berkley DB and LMDB),
-		at the initialization stage we select the metrics that correspond to the selected backend
-	*/
-	switch cfg.Global.BackendImplement {
-	case config.BackendBDB:
-		dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-			"ds_exporter",
-			connPool,
-			"cn=monitor,cn=ldbm database,cn=plugins,cn=config",
-			metrics.GetLdapBDBServerCacheMetrics(),
-			prometheus.Labels{},
-			LdapConnectionPoolTimeout,
-		),
-		)
-		dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-			"ds_exporter",
-			connPool,
-			"cn=database,cn=monitor,cn=ldbm database,cn=plugins,cn=config",
-			metrics.GetLdapBDBDatabaseLDBM(),
-			prometheus.Labels{},
-			LdapConnectionPoolTimeout,
-		),
-		)
-	case config.BackendMDB:
-		dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-			"ds_exporter",
-			connPool,
-			"cn=monitor,cn=ldbm database,cn=plugins,cn=config",
-			metrics.GetLdapMDBServerCacheMetrics(),
-			prometheus.Labels{},
-			LdapConnectionPoolTimeout,
-		),
-		)
-		dsMetricsRegistry.MustRegister(collectors.NewLdapEntryCollector(
-			"ds_exporter",
-			connPool,
-			"cn=database,cn=monitor,cn=ldbm database,cn=plugins,cn=config",
-			metrics.GetLdapMDBDatabaseLDBM(),
-			prometheus.Labels{},
-			LdapConnectionPoolTimeout,
-		),
-		)
-	}
-
-	return dsMetricsRegistry
-}
-
 func run() int {
 	var (
 		applicationResources = appResources{}
@@ -354,7 +96,6 @@ func run() int {
 				BuildTime,
 			),
 		)
-
 		signalCh    = make(chan os.Signal, 1)
 		serverErrCh = make(chan error)
 	)
@@ -362,13 +103,11 @@ func run() int {
 	cfg, err := readConfig(args.ConfigFile)
 	if err != nil {
 		slog.Error("Error loading config", "err", err)
-
 		return 1
 	}
 
 	if args.IsConfigCheck {
 		fmt.Print(cfg.String())
-
 		return 0
 	}
 
@@ -381,7 +120,7 @@ func run() int {
 		)
 
 		defer cancel()
-		err := applicationResources.Shutdown(shutdownContext)
+		err := applicationResources.shutdown(shutdownContext)
 		if err != nil {
 			slog.Error("Shutdown error", "err", err)
 		}
@@ -394,12 +133,11 @@ func run() int {
 		}
 	}()
 
-	logger, logFile, err := setupLogger(cfg)
+	logger, logFile, err := utils.SetupLogger(cfg)
 	applicationResources.LogFile = logFile
 
 	if err != nil {
 		slog.Error("Error initializing logging", "err", err)
-
 		return 1
 	}
 	slog.SetDefault(logger)
@@ -423,16 +161,9 @@ func run() int {
 
 	applicationResources.ConnPool = connections.NewLdapConnectionPool(ldapConnPoolConfig)
 
-	dsMetricsRegistry := setupPrometheusMetrics(cfg, applicationResources.ConnPool)
+	dsMetricsRegistry := metrics.SetupPrometheusMetrics(cfg, applicationResources.ConnPool, LdapConnectionPoolTimeout)
 
-	http.Handle(cfg.HTTP.GetMetricsPath(), promhttp.HandlerFor(dsMetricsRegistry, promhttp.HandlerOpts{}))
-	http.HandleFunc("/", defaultHttpResponse(cfg.HTTP.GetMetricsPath()))
-	http.HandleFunc("/health", healthHttpResponse(applicationResources.ConnPool, startTime))
-	http.HandleFunc("/up", func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte("OK"))
-	})
-
+	// Create HTTP server
 	applicationResources.HttpServer = &http.Server{
 		Addr:         cfg.HTTP.GetListenAddress(),
 		Handler:      http.DefaultServeMux,
@@ -441,14 +172,27 @@ func run() int {
 		IdleTimeout:  time.Duration(cfg.HTTP.GetIdleTimeout()) * time.Second,
 	}
 
+	// Create HTTP Listener with timeouts
 	listener, err := net.Listen("tcp", cfg.HTTP.GetListenAddress())
 	if err != nil {
 		slog.Error("Failed to start TCP listener", "err", err)
-
 		return 1
 	}
-	timeoutListener := connections.NewTimeoutListener(listener, time.Duration(cfg.HTTP.GetInitialReadTimeout())*time.Second)
+	timeoutListener := connections.NewTimeoutListener(
+		listener,
+		time.Duration(cfg.HTTP.GetInitialReadTimeout())*time.Second,
+	)
 
+	// Register HTTP endpoinnts
+	http.Handle(cfg.HTTP.GetMetricsPath(), promhttp.HandlerFor(dsMetricsRegistry, promhttp.HandlerOpts{}))
+	http.HandleFunc("/", utils.DefaultHttpResponse(cfg.HTTP.GetMetricsPath()))
+	http.HandleFunc("/health", utils.HealthHttpResponse(applicationResources.ConnPool, startTime))
+	http.HandleFunc("/up", func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		_, _ = w.Write([]byte("OK"))
+	})
+
+	// Start HTTP server
 	go func() {
 		slog.Info("Starting HTTP server at " + cfg.HTTP.GetListenAddress())
 		err := applicationResources.HttpServer.Serve(timeoutListener)
@@ -470,10 +214,16 @@ func run() int {
 				slog.Info("SIGTERM signal received")
 				running = false
 			case syscall.SIGHUP:
-				slog.Info("SIGHUP signal received - reopening log file")
-				if reopenLogFile(cfg, &applicationResources) != nil {
-					slog.Error("Error reopening log file")
-					running = false
+				slog.Info("SIGHUP signal received")
+				if applicationResources.LogFile != nil {
+					slog.Info("Reopening log file")
+					newLogFile, err := utils.ReopenLogFile(cfg, applicationResources.LogFile)
+					if err != nil {
+						slog.Error("Error reopening log file")
+						running = false
+					}
+					applicationResources.LogFile = newLogFile
+					slog.Info("Log file reopenedd successfully")
 				}
 			}
 		case err := <-serverErrCh:
